services/common/pyproject.toml
[project]
name = "openagent-common"
version = "0.1.0"
description = "Shared primitives (schemas, Protocols, telemetry, logging, DB models) for OpenAgent services"
requires-python = ">=3.13"
dependencies = [
  "pydantic>=2.7",
  "structlog>=24.1",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp-proto-grpc>=1.25",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29",
  "alembic>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openagent_common"]
